% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/aln_merge.R
\name{aln_bwa_merge}
\alias{aln_bwa_merge}
\title{ln_bwa}
\usage{
aln_bwa_merge(fqs1, fqs2, samplename, reflib = getOption("ngs_ref_bwa"),
  bwa_exe = getOption("ngs_bwa_exe"),
  bwa_aln_opts = getOption("ngs_bwa_aln_opts"),
  bwa_sampe_opts = getOption("ngs_bwa_sampe_opts"),
  samtools_exe = getOption("ngs_samtools_exe"), seq_platform = "illumina",
  rgpu = "lane1", center = "MDA")
}
\arguments{
\item{fqs1}{list of fastq files, may be a file with just the fastqs, one in each line.}

\item{fqs2}{list of fastq files, may be a file with just the fastqs, one in each line. mate 2}
}
\description{
ln_bwa
}

