% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate}
\alias{annotate}
\title{Annotate variants using Annovar}
\usage{
annotate(x, samplename = get_opts("samplename"), outfile, build = "hg19",
  annovar_dir = get_opts("annovar_dir"),
  annotate_func = "ngsflows::annotate_mutect")
}
\arguments{
\item{x}{mutect files}

\item{samplename}{}

\item{outfile}{this is the final merged file}
}
\description{
Annotate variants using Annovar
}

