% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_variants}
\alias{annotate_variants}
\title{annotate variants}
\usage{
annotate_variants(x, outpath, outfile, execute = TRUE,
  annovarPath = "/scratch/iacs/apps/annovar/latest",
  db_path = "/scratch/iacs/apps/annovar/humandb", build = "hg19",
  protocols_g, protocols_f, chr = "contig", start = "position",
  end = "position", ref = "ref_allele", alt = "alt_allele",
  use_uuid = TRUE)
}
\arguments{
\item{x}{}

\item{execute}{}

\item{annovarPath}{}

\item{db_path}{}

\item{build}{}

\item{protocols_g}{}

\item{protocols_f}{}

\item{chr}{}

\item{start}{}

\item{end}{}

\item{ref}{}

\item{alt}{}

\item{out_path}{}

\item{out_file}{}
}
\description{
annotate variants
}
\details{
Important default arguments:

\code{-otherinfo -nastring . -remove -buildver}
}

