% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/annotate.R
\name{annotate_vcf}
\alias{annotate_vcf}
\title{Not tested}
\usage{
annotate_vcf(x, oprefix, odir, build, use_uuid, annovarPath, protocols_f,
  protocols_g, ...)
}
\arguments{
\item{x}{}

\item{oprefix}{}

\item{odir}{}

\item{build}{}

\item{use_uuid}{}

\item{annovarPath}{}

\item{protocols_f}{}

\item{protocols_g}{}

\item{...}{}
}
\description{
Not tested
}

