% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch_genomes.R
\name{avail_genomes}
\alias{avail_genomes}
\title{avail_genomes}
\usage{
avail_genomes(species, src, build, from = "igenomes", verbose = FALSE,
  base_url = "ftp://igenome:G3nom3s4u@ussd-ftp.illumina.com", ...)
}
\arguments{
\item{species}{species}

\item{src}{src}

\item{build}{build}

\item{from}{from}

\item{base_url}{}
}
\description{
avail_genomes
}
\examples{
gen = avail_genomes(species = 'Homo_Sapiens', from = 'igenomes')
}

