% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{bam_preprocess}
\alias{bam_preprocess}
\title{Flow following Broad's best practices for variant calling, starting from sorted bam}
\usage{
bam_preprocess(inbam, outbam, q_obj,
  java_exe = "/scratch/rists/hpcapps/x86_64/jdk/jdk1.7.0/bin/java",
  java_tmp = "tmp", cpu_markdup = 4, java_mem_markdup = "-Xmx8g",
  cpu_target = 16, java_mem_target = "-Xmx32g", cpu_realign = 4,
  java_mem_realign = "-Xmx4g", cpu_baserecalib = 4,
  java_mem_baserecalib = "-Xmx4g", cpu_printreads = 4,
  java_mem_printreads = "-Xmx4g",
  gatk_jar = "/scratch/rists/hpcapps/x86_64/gatk/3.1-1/GenomeAnalysisTK.jar",
  picard_dir = "/scratch/rists/hpcapps/x86_64/picard/1.112",
  reffa = "/scratch/rists/hpcapps/reference/human/broad_hg19/fastas/Homo_sapiens_assembly19.fasta",
  gatk_target_opt = "-known /scratch/iacs/iacs_dep/sseth/reference/human/b37/annotations/mills_and_1000g/Mills_and_1000G_gold_standard.indels.b37.vcf -known /scratch/iacs/iacs_dep/sseth/reference/human/b37/annotations/1000g/1000G_phase1.indels.b37.vcf --allow_potentially_misencoded_quality_scores -nt 8",
  gatk_realign_opt = "-known /scratch/iacs/iacs_dep/sseth/reference/human/b37/annotations/mills_and_1000g/Mills_and_1000G_gold_standard.indels.b37.vcf -known /scratch/iacs/iacs_dep/sseth/reference/human/b37/annotations/1000g/1000G_phase1.indels.b37.vcf --allow_potentially_misencoded_quality_scores",
  gatk_baserecalib_opt = "-knownSites /scratch/iacs/iacs_dep/sseth/reference/human/b37/annotations/dbsnp/dbsnp_138.b37.excluding_sites_after_129.vcf -knownSites /scratch/iacs/iacs_dep/sseth/reference/human/b37/annotations/mills_and_1000g/Mills_and_1000G_gold_standard.indels.b37.vcf -knownSites /scratch/iacs/iacs_dep/sseth/reference/human/b37/annotations/1000g/1000G_phase1.indels.b37.vcf --allow_potentially_misencoded_quality_scores -nct 8",
  printreads_opt = "-allowPotentiallyMisencodedQuals -nct 8")
}
\arguments{
\item{inbam}{}

\item{outbam}{}

\item{q_obj}{}

\item{java_exe}{}

\item{java_tmp}{}

\item{cpu_markdup}{}

\item{java_mem_markdup}{}

\item{cpu_target}{}

\item{java_mem_target}{}

\item{cpu_realign}{}

\item{java_mem_realign}{}

\item{cpu_baserecalib}{}

\item{java_mem_baserecalib}{}

\item{cpu_printreads}{}

\item{java_mem_printreads}{}

\item{gatk_jar}{}

\item{picard_dir}{}

\item{reffa}{}

\item{gatk_target_opt}{}

\item{gatk_realign_opt}{}

\item{gatk_baserecalib_opt}{}

\item{printreads_opt}{}
}
\description{
Flow following Broad's best practices for variant calling, starting from sorted bam
}

