% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bwa.R
\name{bwa}
\alias{bwa}
\title{Wrapper for BWA sequence alignment tool}
\usage{
bwa(fqs1, fqs2, samplename = get_opts("samplename"), paired_end,
  bwa_exe = get_opts("bwa_exe"), ref_bwa = get_opts("ref_bwa"),
  bwa_aln_opts = get_opts("bwa_aln_opts"),
  cpu_bwa_aln = get_opts("cpu_bwa_aln"),
  bwa_sampe_opts = get_opts("bwa_sampe_opts"),
  bwa_samse_opts = get_opts("bwa_samse_opts"),
  samtools_exe = get_opts("samtools_exe"))
}
\arguments{
\item{fastq1}{character vector with full paths to fastq files for mate 1.}

\item{fastq2}{character vector}
}
\value{
A list, with cmds as one of the variables which
contains system commands to be run.
}
\description{
Wrapper for BWA sequence alignment tool
}
\details{
Default params picked up from ngsflow.conf
}

