% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sheets_fastqs.R
\name{check_fastq_sheet}
\alias{check_fastq_sheet}
\title{check_fastq_sheet}
\usage{
check_fastq_sheet(mat, id_column = "sample_id", file_column = "files",
  read_column = "read")
}
\arguments{
\item{mat}{}
}
\description{
check_fastq_sheet
}

