% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sheets_fastqs.R
\name{create_fq_sheet}
\alias{create_fq_sheet}
\title{create a sheet of fastq}
\usage{
create_fq_sheet(x, ext = get_opts("ngs_fq_ext"),
  format = get_opts("ngs_fq_format"), sample_prefix = "")
}
\arguments{
\item{x}{path to a fastq folder}

\item{ext}{extensions to look for. A regex to search for fastq files}

\item{format}{auto detect. Pattern specify pattern acceptable to split_fastq_names. If missing will detect hiseq and miseq}

\item{sample_prefix}{A prefix to add to all sample names, run, project etc.....}
}
\description{
create a sheet of fastq
}

