% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sheets_paired.R
\name{create_paired_samplesheet}
\alias{create_paired_samplesheet}
\title{create_paired_samplesheet}
\usage{
create_paired_samplesheet(x, fqmat, sampbam, refbam, outfile, db_sampleid = 0,
  db_refid = 0, out_prefix, bampath, project = "project",
  subproject = "subproject",
  normal_bam_01 = "/scratch/iacs/ngs/commonNormalBam/ANDY-Hetero-JZ-330-10-01_130425_SN208_0465_BD2705ACXX_merged2_rg.sorted.recalibed.bam")
}
\arguments{
\item{x}{}

\item{fqmat}{}

\item{sampbam}{}

\item{refbam}{}

\item{outfile}{}

\item{db_sampleid}{}

\item{db_refid}{}

\item{out_prefix}{}

\item{bampath}{}

\item{project}{}

\item{normal_bam_01}{}
}
\description{
create_paired_samplesheet
}

