% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sheets_paired.R
\name{create_tooling_paired_samplesheet}
\alias{create_tooling_paired_samplesheet}
\title{create_tooling_paired_samplesheet}
\usage{
create_tooling_paired_samplesheet(x, outfile, tumor.only = FALSE,
  normal.only = FALSE)
}
\arguments{
\item{x}{}

\item{outfile}{}

\item{tumor.only}{}

\item{normal.only}{}
}
\description{
create_tooling_paired_samplesheet
}

