% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastqc.R
\name{fastqc}
\alias{fastqc}
\title{Fastqc}
\usage{
fastqc(fqs, fqpath, odir, fastqc_exe = get_opts("fastqc_exe"),
  cpu_fastqc = get_opts("cpu_fastqc"), casava = TRUE,
  fastqc_opts = get_opts("fastqc_opts"))
}
\arguments{
\item{fqs}{}

\item{fqpath}{}

\item{odir}{}

\item{fastqc_exe}{}

\item{cpu_fastqc}{}

\item{casava}{}

\item{fastqc_opts}{}
}
\description{
Fastqc
}

