% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fetch_genomes.R
\name{fetch_genomes}
\alias{fetch_genomes}
\title{fetch_genomes}
\usage{
fetch_genomes(genome_path = "~/flowr/genomes", verbose = FALSE, ...)
}
\arguments{
\item{...}{passed onto \link{fetch_genomes}}

\item{base_url}{}

\item{species}{}

\item{src}{}

\item{build}{}
}
\description{
fetch_genomes
}
\examples{
\\donotrun{
genomes_fetch(species = Homo_sapiens, src = 'NCBI', build = 'build37.2')
}
}

