% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_genomes.R
\name{genomes_avail}
\alias{genomes_avail}
\title{genomes_avail}
\usage{
genomes_avail(species, src, build, from = "igenomes", verbose = FALSE,
  base_url = "ftp://igenome:G3nom3s4u@ussd-ftp.illumina.com", ...)
}
\arguments{
\item{species}{species}

\item{src}{src}

\item{build}{build}

\item{from}{from}

\item{base_url}{}
}
\description{
genomes_avail
}
\examples{
gen = genomes_avail(species = 'Homo_Sapiens', from = 'igenomes')
}

