% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/fetch_genomes.R
\name{genomes_fetch}
\alias{genomes_fetch}
\title{genomes_fetch}
\usage{
genomes_fetch(genome_path = "~/flowr/genomes", ...)
}
\arguments{
\item{...}{passed onto \link{genomes_fetch}}

\item{base_url}{}

\item{species}{}

\item{src}{}

\item{build}{}
}
\description{
genomes_fetch
}
\examples{
\\donotrun{
genomes_fetch(species = Homo_sapiens, src = 'NCBI', build = 'build37.2')
}
}

