% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/preprocess.R
\name{get_fasta_chrs}
\alias{get_fasta_chrs}
\title{Read the associated dictionary file and return a list of chromosome names}
\usage{
get_fasta_chrs(x)
}
\arguments{
\item{x}{a reference genome fasta file}
}
\description{
Read the associated dictionary file and return a list of chromosome names
}

