% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/haplotyper.R
\name{haplotyper}
\alias{haplotyper}
\title{a wrapper around gatk haplotyper}
\usage{
haplotyper(x, samplename = get_opts("samplename"), outfile,
  split_by_chr = TRUE, ref_fasta = get_opts("ref_fasta"),
  java_exe = get_opts("java_exe"), java_mem = get_opts("java_mem"),
  java_tmp = get_opts("java_tmp"), gatk_jar = get_opts("gatk_jar"),
  haplotyper_opts = get_opts("haplotyper_opts"),
  cpu_haplotyper = get_opts("cpu_haplotyper"))
}
\arguments{
\item{x}{input bam}
}
\description{
a wrapper around gatk haplotyper
}
\details{
For more details refer to: https://www.broadinstitute.org/gatk/gatkdocs/org_broadinstitute_gatk_tools_CatVariants.php
}

