% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mutect.R
\name{mutect}
\alias{mutect}
\title{Title}
\usage{
mutect(tumor_bam, normal_bam, samplename = get_opts("samplename"), outfile,
  is_merged = TRUE, split_by_chr = TRUE, java_exe = get_opts("java_exe"),
  java_tmp = get_opts("java_tmp"), mutect_jar = get_opts("mutect_jar"),
  cpu_mutect = get_opts("cpu_mutect"), mem_mutect = get_opts("java_mem"),
  ref_fasta = get_opts("ref_fasta"), mutect_opts = get_opts("mutect_opts"))
}
\arguments{
\item{tumor_bam}{}

\item{normal_bam}{}

\item{samplename}{}

\item{outfile}{}

\item{is_merged}{}

\item{split_by_chr}{}
}
\description{
Title
}
\examples{
\dontrun{

x = "tumor.bam"
y = "normal.bam"

out = mutect(x, y, is_merged = TRUE)

}
}

