% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/picard.R
\name{picard_reorder}
\alias{picard_reorder}
\title{Use picard's MergeSamFiles tool to merge bam/sam files}
\usage{
picard_reorder(x, samplename = get_opts("samplename"),
  java_exe = get_opts("java_exe"), java_mem = get_opts("java_mem"),
  java_tmp = get_opts("java_tmp"), picard_dir = get_opts("picard_dir"),
  picard_reorder_opts = get_opts("picard_reorder_opts"))
}
\arguments{
\item{x}{a vectors of files to merge}

\item{samplename}{}

\item{java_exe}{}

\item{java_mem}{}

\item{java_tmp}{}

\item{picard_dir}{}

\item{mergedbam}{}
}
\description{
The resulting file is sorted and index is created for it.
Validation stringency of inputs is kept as lenient.
Multi-threading is turned on by default, though in our experience this does
not seem to use a lot of threads.
}

