% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/picard.R
\name{picard_rg}
\alias{picard_rg}
\title{picard_rg}
\usage{
picard_rg(x, samplename = get_opts("samplename"), lane = "lane1",
  seq_platform = "illumina", center = "MDA",
  java_exe = get_opts("java_exe"), java_mem = get_opts("java_mem"),
  java_tmp = get_opts("java_tmp"), picard_dir = get_opts("picard_dir"))
}
\description{
picard_rg
}

