% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sheets_paired.R
\name{read_paired_samplesheet}
\alias{read_paired_samplesheet}
\title{read_paired_samplesheet}
\usage{
read_paired_samplesheet(x, ...)
}
\arguments{
\item{x}{}

\item{...}{}
}
\description{
read_paired_samplesheet
}

