% Generated by roxygen2 (4.0.1.99): do not edit by hand
\name{split.names.fastq}
\alias{split.names.fastq}
\title{split.names.fastq}
\usage{
\method{split}{names.fastq}(files,
  format = "$samplename$_$index$_L00$lane$_R$read$_$num$.fastq.gz")
}
\arguments{
\item{files}{}

\item{format}{}
}
\description{
split.names.fastq
given a format split the files provided. Tried and tested on fastq files
}

