% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mod_split_fq.R, R/split_fq.R
\name{split_fq}
\alias{split_fq}
\title{split_fq}
\usage{
split_fq(x, samplename, jobname,
  split_fq_exe = system.file("scripts/split_fq.sh", package = "ngsflows"),
  chunks = 200)

split_fq(x, samplename, jobname,
  split_fq_exe = system.file("scripts/split_fq.sh", package = "ngsflows"),
  chunks = 200)
}
\arguments{
\item{x}{file to split}

\item{chunks}{approx size of resulting file, in MB}

\item{split_size}{approx size of resulting file, in MB}

\item{x}{file to split}
}
\description{
split_fq

split_fq
}
\details{
estimate the number of splits required
pre-compute the number of splits it would generate
make commands to make the split
approx, make 200mb files, for splitting...

estimate the number of splits required
pre-compute the number of splits it would generate
make commands to make the split
approx, make 200mb files, for splitting...
}

