% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/star.R
\name{star.2.4.2a}
\alias{star}
\alias{star.2.4.2a}
\title{A flowr wrapper for STAR aligner}
\usage{
star.2.4.2a

star(fq1, fq2, bam,
 samplename = "samp1", out_path, out_prefix,
 ref_fasta = get_opts("ref_fasta"), ref_star = get_opts("ref_star"),
 star_exe = get_opts("star_exe"), star_opts = get_opts("star_opts"),
 star_cpu = get_opts("star_cpu"))
}
\arguments{
\item{fq1}{single fastq file}

\item{fq2}{}

\item{bam}{alternatively use of bam as input, instead of fqs}

\item{samplename}{}

\item{out_path}{}

\item{out_prefix}{}

\item{ref_fasta}{}

\item{ref_star}{}

\item{star_exe}{}

\item{star_opts}{}

\item{star_cpu}{}
}
\description{
For more details refer to \href{https://github.com/alexdobin/STAR/raw/master/doc/STARmanual.pdf}{STAR's manual}.
Starting 2.4.2a, star supports a twoPass mode, thus simplifying this pipeline.
}

