% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star.R
\name{star_index}
\alias{star_index}
\title{star_index}
\usage{
star_index(ref_star = get_opts("ref_star"), star_exe = get_opts("star_exe"),
  ref_fasta = get_opts("ref_fasta"), star_cpu = get_opts("star_cpu"),
  ref_gtf = get_opts("ref_gtf"))
}
\description{
a helper function to create a index, incase it does not exits
}

