% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bam_preprocess.R
\name{bam_preprocess}
\alias{bam_preprocess}
\title{Flow following Broad's best practices for variant calling, starting from sorted bam}
\usage{
bam_preprocess(inbam, outbam, samplename = "samp", split_by_chr = FALSE,
  java_exe = get_opts("java_exe"), java_tmp = get_opts("java_tmp"),
  cpu_markdup = 4, mem_markdup = "-Xmx8g", cpu_target = 16,
  mem_target = "-Xmx32g", cpu_realign = 8, mem_realign = "-Xmx4g",
  cpu_baserecalib = 4, mem_baserecalib = "-Xmx4g", cpu_printreads = 4,
  mem_printreads = "-Xmx4g", gatk_jar = get_opts("gatk_jar"),
  picard_dir = get_opts("picard_dir"), ref_fasta = get_opts("ref_fasta"),
  gatk_target_opts = get_opts("gatk_target_opts"),
  gatk_realign_opts = get_opts("gatk_baserecalib_opts"),
  gatk_baserecalib_opts = get_opts("gatk_baserecalib_opts"),
  gatk_printreads_opts = get_opts("gatk_target_opts"))
}
\arguments{
\item{inbam}{}

\item{outbam}{}

\item{java_exe}{path to java}

\item{java_tmp}{path to java tmp, can leave blank}

\item{cpu_markdup}{cpus to be reserved, default 4}

\item{cpu_target}{}

\item{cpu_realign}{}

\item{cpu_baserecalib}{}

\item{cpu_printreads}{}

\item{gatk_jar}{}

\item{picard_dir}{}

\item{q_obj}{is provided output is a flow object, else a list of commands to run}

\item{java_mem_markdup}{memory provided to java}

\item{java_mem_target}{}

\item{java_mem_realign}{}

\item{java_mem_baserecalib}{}

\item{java_mem_printreads}{}

\item{reffa}{}

\item{gatk_target_opt}{}

\item{gatk_realign_opt}{}

\item{gatk_baserecalib_opt}{}

\item{printreads_opt}{}
}
\description{
This function provides a list of commands to run.
CPU's refer to number of cores reserved the job. We have a seperate param to control the number of threads the command would use. Example. On a node with 10GB and 10 cores. We may only reserve 10, but use only 2 threads for a 5GB memory command.
}
\details{
Flow following Broad's best practices for variant calling, starting from sorted bam
}

